requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
Pillow==10.1.0
plotly==5.17.0
python-dotenv==1.0.0
